pyproject.toml
[project]
name = "library-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.1",
    "asyncpg>=0.30.0",
    "bcrypt==3.2.2",
    "fastapi[standard]>=0.115.12",
    "greenlet>=3.2.2",
    "passlib>=1.7.4",
    "pydantic>=2.11.5",
    "pydantic-settings>=2.9.1",
    "python-jose>=3.5.0",
    "ruff>=0.11.12",
    "sqlalchemy>=2.0.41",
    "uuid>=1.30",
    "uvicorn>=0.34.3",
]
